+++ src/main.ts
@@ -2,7 +2,8 @@ import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
-import { environment, fetchConfigPromise } from './environments/environment';
+import { environment } from './environments/environment';
+import { fetchConfigPromise } from './environments/environment.types';
 
 const copyFields = <Type>(source: Type, target: Type) => {
   for (const key in source) {
